Update AppService_FunctionApp_Audit_java_Latest.json
Field `Microsoft.Web/sites/config/web.javaVersion` outputs java version as `1.x` where `x` is the Java version used. The current concat adds an asterisk at the end of the statement and thus does not match the output. Confirmed in our subscriptions moving the asterisk to the beginning of the match allows the policy to correctly identify Java version for function apps.