README Update: Adding known issue for resource types with wrong non-compliance message
There is a known list of resources and RPs that do not support the non-compliance message feature and instead show links to the activity log. This PR adds this list in the known-issue section of the README file. For more info see CSS support case 2107010060003670 and corresponding IcMs & Collab cases with PG teams about this feature.